+++ run_tests.py
@@ -1,11 +1,11 @@
-import tests.problems as ivps
+import unit_tests.problems as ivps
 
 from csunisa.lmm import LinearMultistepMethod
 from csunisa.lmm_registry import get_method, list_methods
-from tests.tester import test_lmm
+from unit_tests.tester import test_lmm
 
 
-def unit_test_lmms(h=0.05):
+def unit_test_lmms(h=0.01):
     methods = list_methods()
 
     for method in methods: